backend/requirements.txt
uvicorn==0.34.3
fastapi==0.115.13
requests==2.32.4
google-generativeai==0.8.5
python-dotenv==1.1.0
httpx==0.28.1
pydantic-ai==0.2.4